minor refactor move each class to its own file
