+++ client/src/main/scala/io/delta/sharing/client/auth/AuthCredentialProviderFactory.scala
@@ -1,0 +1,33 @@
+/*
+ * Copyright (2021) The Delta Lake Project Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.delta.sharing.client.auth
+
+import org.apache.http.impl.client.CloseableHttpClient
+
+import io.delta.sharing.client.{BearerTokenDeltaSharingProfile, DeltaSharingProfile, OAuthClientCredentialsDeltaSharingProfile}
+
+private[client] object AuthCredentialProviderFactory {
+  def createCredentialProvider(profile: DeltaSharingProfile,
+                               client: CloseableHttpClient): AuthCredentialProvider = {
+    profile match {
+      case oauthProfile: OAuthClientCredentialsDeltaSharingProfile =>
+        OAuthClientCredentialsAuthProvider(client, oauthProfile)
+      case BearerTokenDeltaSharingProfile(_, _, bearerToken, _) =>
+        BearerTokenAuthProvider(bearerToken)
+    }
+  }
+}
